Moves async code
