+++ tests/cacheables/asyncio/caches/test_disk.py
@@ -1,7 +1,7 @@
 import asyncio
 import os
 import pytest
-from cacheables_async.caches.disk import AsyncDiskCache
+from cacheables.asyncio.caches.disk import AsyncDiskCache
 from cacheables.exceptions import ReadException, InputKeyNotFoundError
 from cacheables.keys import FunctionKey, InputKey
